+++ app/system/app/components/modal.vue
@@ -1,30 +1,27 @@
 <template>
     <div uk-modal :class="modalCls">
         <div class="uk-modal-dialog" :class="dialogCls">
-            <slot v-if="opened">
-            </slot>
+            <slot v-if="opened"> </slot>
         </div>
     </div>
 </template>
 
 <script>
-
-import { on } from 'uikit-util';
+import { on } from "uikit-util";
 
 export default {
-
     props: {
         large: Boolean,
         lightbox: Boolean,
         modalFull: Boolean,
         modalSmall: Boolean,
-        widthAuto : Boolean,
-        center    : Boolean,
-        contrast  : Boolean,
-        bgClose   : Boolean,
-        escClose  : Boolean,
-        closed    : Function,
-        modifier: { type: String, default: '' },
+        widthAuto: Boolean,
+        center: Boolean,
+        contrast: Boolean,
+        bgClose: Boolean,
+        escClose: Boolean,
+        closed: Function,
+        modifier: { type: String, default: "" },
         options: {
             type: Object,
             default() {
@@ -40,20 +37,19 @@ export default {
     },
 
     computed: {
-
         modalCls() {
-            const modalCls = this.modifier.split(' ');
+            const modalCls = this.modifier.split(" ");
 
             if (this.large) {
-                modalCls.push('uk-modal-container');
+                modalCls.push("uk-modal-container");
             }
 
             if (this.modalFull) {
-                modalCls.push('uk-modal-full');
+                modalCls.push("uk-modal-full");
             }
 
             if (this.modalSmall) {
-                modalCls.push('tm-modal-small');
+                modalCls.push("tm-modal-small");
             }
 
             return modalCls;
@@ -63,32 +59,37 @@ export default {
             const dialogCls = [];
 
             if (this.center) {
-                dialogCls.push('uk-margin-auto-vertical');
+                dialogCls.push("uk-margin-auto-vertical");
             }
 
             if (this.contrast) {
-                dialogCls.push('uk-light');
+                dialogCls.push("uk-light");
             }
 
             if (this.widthAuto) {
-                dialogCls.push('uk-width-auto');
+                dialogCls.push("uk-width-auto");
             }
 
             return dialogCls;
         },
-
     },
 
     mounted() {
         const vm = this;
 
-        this.modal = UIkit.modal(this.$el, _.extend({
-            bgClose: !this.bgClose,
-            escClose: !this.escClose,
-            stack: true,
-        }, this.options));
+        this.modal = UIkit.modal(
+            this.$el,
+            _.extend(
+                {
+                    bgClose: !this.bgClose,
+                    escClose: !this.escClose,
+                    stack: true,
+                },
+                this.options
+            )
+        );
 
-        on(this.modal.$el, 'hidden', () => {
+        on(this.modal.$el, "hidden", () => {
             vm.opened = false;
             if (this.closed) {
                 this.closed();
@@ -97,7 +98,6 @@ export default {
     },
 
     methods: {
-
         open() {
             this.opened = true;
             this.modal.show();
@@ -106,9 +106,6 @@ export default {
         close() {
             this.modal.hide();
         },
-
     },
-
 };
-
 </script>